// portal/proto/isuxportal/services/admin/clarifications_pb.d.ts
export declare type CreateClarificationRequest = Message<"isuxportal.proto.services.admin.CreateClarificationRequest"> & {
  /**
   * @generated from field: string answer = 2;
   */
  answer: string;

  /**
   * @generated from field: string question = 3;
   */
  question: string;

  /**
   * @generated from field: int64 team_id = 4;
   */
  teamId: bigint;
};

/**
 * Describes the message isuxportal.proto.services.admin.CreateClarificationRequest.
 * Use `create(CreateClarificationRequestSchema)` to create a new message.
 */
export declare const CreateClarificationRequestSchema: GenMessage<CreateClarificationRequest>;